Lift RawGeotag's CR3, GPX and XMP engine into crates/geotag
Decision 17 takes three hard sub-problems as already solved -- CR3 capture
time by seeking the container rather than reading 45 MB, GPX indexing with the
gap and distance refusals, and XMP packets diffed against Lightroom Classic
15.4.1 to 0.02-0.12 m. They are solved because they were checked against
thousands of real frames on two bodies, so the lift moved them unrewritten.
Tidying them up in transit would have spent exactly what the decision was
trying to save. All 67 of their unit tests came across and pass.

Two changes were made, both because photoday cannot do without them:

  - raw::capture_time_in_memory. Phase 3 holds every file in RAM to hash it
    (decision 10), so re-reading it from the card to find its capture time is
    the waste that decision exists to avoid. It shares the timezone policy with
    the path-based function rather than copying it -- that is the part that can
    misplace a photo by a day of travel, and it now has one body.

    It still consults read_strategy, which looks redundant and is not. The
    strategy records which parser path a format survives, not how much of the
    file to read: NEF fails through the seeking path and parses only through
    the memory one. So CR3 gets a cursor over the caller's slice and copies
    nothing, and NEF pays one copy because that is the only path it parses on.

  - xmp::render takes the writing tool's identity for x:xmptk instead of baking
    in "rawgeotag". Two tools emit these packets now, and a sidecar naming the
    wrong one is a small lie in a file whose whole job is provenance.

Verified against real frames, since no committed test can reach the in-memory
success path without a fixture: it agrees with the path-based function on a
+01:00 CR3 (the BST case decision 23 describes), a +00:00 CR3, and a NEF with
no offset at all. The new xmptk test was mutation-checked -- hard-coding the
writer again fails it and the byte-for-byte packet test, and nothing else.

Also found by using the API rather than reading it: geotag now re-exports
MediaParser, which both entry points take. Without that every consumer needed
its own nom-exif dependency purely to spell an argument, and could then drift
onto a different version than the one this crate parses with.

Decision 29 is corrected in two places, both discovered by doing the work:

  - It claimed two thiserror sites. There is one. raw::Capture is already an
    enum of outcomes, so the distinction decision 21 needs was solved before
    this design asked for it, and the in-memory path closes the rest of the
    gap: with the bytes already in RAM there is no I/O failure left to tell
    from a defective file. A design that reasons about code it has not read
    will invent error types the code already made unnecessary.

  - LIGHTROOM-XMP.md was copied here and then removed. Its procedure does not
    describe the engine, it drives rawgeotag.exe -- staging real frames and
    diffing sidecars. photoday cannot write a sidecar yet, so the document's
    first instruction would be unrunnable in the repository holding it, which
    is the exact failure WRITING.md opens with. UPDATING.md keeps pointing at
    RawGeotag; the move waits for phase 5.

The lift leaves a deliberate duplication. RawGeotag was not modified and still
builds and runs as before, so the engine now exists twice and a fix to one copy
does not reach the other. Resolving it means changing that repository -- a path
dependency or a retirement -- and that is not this repo's call to make.

photoday now consumes the engine rather than merely depending on it:
--max-gap-seconds and --max-gap-meters take their defaults from
GapLimits::DEFAULT, so the limits the CLI advertises and the limits
Track::lookup enforces cannot drift apart.

+++ crates/photoday/src/main.rs
@@ -10,6 +10,7 @@ use std::path::PathBuf;
 use std::process::ExitCode;
 
 use clap::{Args, Parser, Subcommand};
+use geotag::track::GapLimits;
 
 #[derive(Debug, Parser)]
 #[command(
@@ -71,12 +72,17 @@ struct Offload {
     no_gpx: bool,
 
     /// Refuse to interpolate across a longer hole.
-    #[arg(long, value_name = "S", default_value_t = 60)]
-    max_gap_seconds: u64,
+    //
+    // Both defaults come from the engine rather than being written again here, so the
+    // limit the CLI advertises and the limit `Track::lookup` enforces cannot drift
+    // apart. Decision 16 renamed these from RawGeotag's `--max-gap`/`--max-distance`;
+    // the values behind them are the same ones, and deliberately harsh.
+    #[arg(long, value_name = "S", default_value_t = GapLimits::DEFAULT_GAP_SECONDS)]
+    max_gap_seconds: i64,
 
     /// Refuse to interpolate across a wider hole.
-    #[arg(long, value_name = "M", default_value_t = 100)]
-    max_gap_meters: u64,
+    #[arg(long, value_name = "M", default_value_t = GapLimits::DEFAULT.max_meters)]
+    max_gap_meters: f64,
 
     /// Overwrite existing XMP; archives only unless a destination is named.
     //
